fixup! KTOR-1647 Extract autoreload in the separate module
