+++ ktor-server/ktor-server-plugins/ktor-server-autoreload/jvm/src/io/ktor/server/autoreload/ApplicationEngineEnvironmentReloading.kt
@@ -2,14 +2,17 @@
  * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.server.engine
+package io.ktor.server.autoreload
 
 import io.ktor.events.*
 import io.ktor.events.EventDefinition
 import io.ktor.http.*
 import io.ktor.server.application.*
+import io.ktor.server.autoreload.internal.*
+import io.ktor.server.autoreload.internal.currentStartupModules
+import io.ktor.server.autoreload.internal.get_com_sun_nio_file_SensitivityWatchEventModifier_HIGH
 import io.ktor.server.config.*
-import io.ktor.server.engine.internal.*
+import io.ktor.server.engine.*
 import io.ktor.util.*
 import io.ktor.util.logging.*
 import io.ktor.util.pipeline.*
