+++ ktor-server/ktor-server-plugins/ktor-server-autoreload/jvm/src/io/ktor/server/autoreload/internal/CallableUtils.kt
@@ -1,8 +1,8 @@
 /*
- * Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.server.engine.internal
+package io.ktor.server.autoreload.internal
 
 import io.ktor.server.application.*
 import java.lang.reflect.*
