+++ ktor-server/ktor-server-plugins/ktor-server-autoreload/jvm/src/io/ktor/server/autoreload/internal/OverridingClassLoader.kt
@@ -2,7 +2,7 @@
  * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
  */
 
-package io.ktor.server.engine.internal
+package io.ktor.server.autoreload.internal
 
 import java.io.*
 import java.net.*
