+++ ktor-server/ktor-server-plugins/ktor-server-autoreload/jvm/src/io/ktor/server/autoreload/internal/ServerHostUtils.kt
@@ -1,8 +1,8 @@
 /*
-* Copyright 2014-2021 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
-*/
+ * Copyright 2014-2022 JetBrains s.r.o and contributors. Use of this source code is governed by the Apache 2.0 license.
+ */
 
-package io.ktor.server.engine
+package io.ktor.server.autoreload.internal
 
 import kotlin.reflect.*
 import kotlin.reflect.jvm.*
